Move timeouts from Async to Api module
This will allow resoruces that aren't explicitly async to have a configurable
timeout. This also adds a default timeout for all api calls made by Terraform.
