+++ templates/global_async.yaml.erb
@@ -5,7 +5,7 @@ async: !ruby/object:Api::Async
     base_url: 'projects/{{project}}/global/operations/{{op_id}}'
     wait_ms: 1000
 <% unless ctx[:timeouts].nil? -%>
-    timeouts: !ruby/object:Api::Async::Operation::Timeouts
+    timeouts: !ruby/object:Api::Timeouts
 <%= lines(indent(ctx[:timeouts].map { |k, v| "#{k}: #{v}" }, 6)) -%>
 <% end -%>
   result: !ruby/object:Api::Async::Result
