+++ templates/terraform/resource.html.markdown.erb
@@ -38,6 +38,7 @@
 <%
   resource_name = product_ns.underscore.downcase + '_' + object.name.underscore
   properties = object.all_user_properties
+  timeouts = object.async.nil? ? Api::Timeouts.new : object.async.operation.timeouts
 -%>
 ---
 <%= lines(autogen_notice :yaml) -%>
@@ -147,18 +148,16 @@ In addition to the arguments listed above, the following computed attributes are
 <% end -%>
 <% end -%>
 
-<% unless object.async.nil? -%>
 ## Timeouts
 
 This resource provides the following
 [Timeouts](/docs/configuration/resources.html#timeouts) configuration options:
 
-- `create` - Default is <%= object.async.operation.timeouts.insert_sec / 60 -%> minutes.
+- `create` - Default is <%= timeouts.insert_sec / 60 -%> minutes.
 <% if updatable?(object, properties) -%>
-- `update` - Default is <%= object.async.operation.timeouts.update_sec / 60 -%> minutes.
+- `update` - Default is <%= timeouts.update_sec / 60 -%> minutes.
 <% end -%>
-- `delete` - Default is <%= object.async.operation.timeouts.delete_sec / 60 -%> minutes.
-<% end -%>
+- `delete` - Default is <%= timeouts.delete_sec / 60 -%> minutes.
 
 ## Import
